Modularizes the actor list and the add actor input to clean up the home component. Also, rewrites the addActor component to use onChange to change a piece of state instead of using useRef to track it.
